# [VLM] Avoid unnecessary dummy multimodal data during processing