Currently with the cache enabled, in `BaseMultiModalProcessor._apply_hf_processor_mm_only` we create both text and multimodal dummy data but the latter is unused, leading to unnecessary overhead for allocating NumPy arrays. This is especially devastating for models that support large image sizes such as Qwen2-VL series.

This PR optimizes this process by splitting up the dummy data generation into text and multimodal parts, so that we can create just the dummy text in `BaseMultiModalProcessor._apply_hf_processor_mm_only`.

**Note:** Multi-modal processors will be required to implement `get_dummy_text` and `get_dummy_mm_data` in a future release. This PR adds a warning asking the developer to do so if the processor hasn't done that yet.

Benchmark (#11196):
```
python benchmarks/mmmu_bench.py --model Qwen/Qwen2.5-VL-3B-Instruct --max-model-len 16384 --max-num-seqs 64 --num-prompts 250
```

Main branch
```
# Trial 1
Request throughput: 0.36 req/s
Total generated tokens: 1149
Token generation rate: 1.66 tok/s

# Trial 2
Request throughput: 0.51 req/s
Total generated tokens: 1151
Token generation rate: 2.33 tok/s

# Trial 3
Request throughput: 0.63 req/s
Total generated tokens: 1180
Token generation rate: 3.00 tok/s
```

This branch
```
# Trial 1
Request throughput: 0.54 req/s
Total generated tokens: 1173
Token generation rate: 2.51 tok/s

# Trial 2
Request throughput: 0.54 req/s
Total generated tokens: 1138
Token generation rate: 2.44 tok/s

# Trial 3
Request throughput: 0.57 req/s
Total generated tokens: 1191
Token generation rate: 2.74 tok/s
```